Fixed changing twab bug in TwabController

+++ test/invariants/handlers/TwabControllerHandler.sol
@@ -279,9 +279,9 @@ contract TwabControllerHandler is CommonBase, StdCheats, StdUtils {
         }
 
         // Check if each period end timestamp is safe for this vault
-        uint48 newestPeriod = twabController.getTimestampPeriod(newestObservation.timestamp);
-        for (uint48 p_i = 1; p_i < newestPeriod; ++p_i) {
-          uint48 timestamp = PERIOD_OFFSET + (p_i * PERIOD_LENGTH);
+        uint32 newestPeriod = twabController.getTimestampPeriod(newestObservation.timestamp);
+        for (uint32 p_i = 1; p_i < newestPeriod; ++p_i) {
+          uint32 timestamp = PERIOD_OFFSET + (p_i * PERIOD_LENGTH);
           isVaultsSafe = isVaultsSafe && twabController.hasFinalized(timestamp);
         }
